【flutter】【demo】【flutter】【demo】Sample Demo code structure, logic optimization
