+++ TRTC-Simple-Demo/lib/main.dart
@@ -1,13 +1,12 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:provider/provider.dart';
-import 'package:trtc_demo/page/trtcmeetingdemo/index.dart';
-import 'package:trtc_demo/page/trtcmeetingdemo/meeting.dart';
-import 'package:trtc_demo/page/trtcmeetingdemo/member_list.dart';
-import 'package:trtc_demo/page/trtcmeetingdemo/test_api.dart';
-import 'package:trtc_demo/models/meeting.dart';
-import 'package:trtc_demo/page/trtcmeetingdemo/texture_render.dart';
-import 'package:trtc_demo/page/trtcmeetingdemo/test_web.dart';
+import 'package:trtc_demo/ui/login.dart';
+import 'package:trtc_demo/ui/meeting.dart';
+import 'package:trtc_demo/ui/member_list.dart';
+import 'package:trtc_demo/ui/test/test_api.dart';
+import 'package:trtc_demo/models/meeting_model.dart';
+import 'package:trtc_demo/ui/test/test_web.dart';
 
 void main() {
   WidgetsFlutterBinding.ensureInitialized();
@@ -29,10 +28,9 @@ class _MyAppState extends State<MyApp> {
       create: (context) => MeetingModel(),
       child: MaterialApp(
         routes: {
-          "/": (context) => IndexPage(),
-          "/index": (context) => IndexPage(),
-          "/video": (context) => MeetingPage(),
-          "/textureRender": (context) => TextureRenderPage(),
+          "/": (context) => LoginPage(),
+          "/login": (context) => LoginPage(),
+          "/meeting": (context) => MeetingPage(),
           "/memberList": (context) => MemberListPage(),
           "/test": (context) => TestPage(),
           "/testweb": (context) => TestWebPage()
