+++ TRTC-Simple-Demo/lib/models/data_models.dart
@@ -1,0 +1,14 @@
+
+class WidgetSize {
+  int width;
+  int height;
+
+  WidgetSize({required int this.width, required int this.height});
+}
+
+enum BeautyType {
+  smooth,
+  nature,
+  pitu,
+  ruddy,
+}
